converted main.rs to lib.rs to be exported as a library crate
